[SF#502437] rogue reactor and unittest
in short, the nosy reactor was modifying the nosy list. That code had
been there for a long time, and I suspect it was there because we
weren't generating the nosy list correctly in other places of the code.
We're now doing that, so the nosy-modifying code can go away from the
nosy reactor.