factor getCurrentURL into its own method:
I'm using a query on a class with two different templates. Now I want
to be able to store the template as part of the query. The current
indexargs_url implementation strips the @template part.

Factoring the URL computation lets me override the action to use in my
template. But maybe the url-computation should be changed to include the
template in use when storing a query? See example code in the comment.