Fix [SF#738470]. Validate timezone setting in userauditor
The timezone setting is just a plain string, which was being
stored blindly. The problem is that certain parts of the roundup
web UI (notably the user details page) try to use, assuming that it
is valid. If a user enters garbage then they cannot access their
details page (and possibly other parts of roundup) until an admin
fixes their timezone for them.

During userauditor, we attempt to use the new timezone value and
catch any exceptions that may be raised.

I posted this patch to roundup-users several days ago and no one
screamed. Now that I have commit access I'll check it in.

I also noticed that the classic and minimal template userauditors
were slightly different in their check for 'roles'. I couldn't discern
a reason for the difference so I made them the same. I can change it
back if need be.