Add note about dropping support for Python v2.5
After some discussion on the roundup-devel mailing list it was agreed
that support for Python v2.5 should be dropped in the v1.6 release of
Roundup.

Dropping support for Python v2.5 should allow a number of cleanups to
the code base as well a provide and opportunity to start porting Roundup
to Python v3.