issue2550850 anypy/email_.py uses BSPACE which is not defined in python 2.7
Supplied a definition for BSPACE since it seems to not be defined
anywhere (even searching a python 3.4 install didn't turn up a definition for
it). BSPACE definition taken from:

  http://svn.python.org/projects/python/branches/py3k-urllib/Lib/email/header.py

which is almost identical code. Put definition of BSPACE just before
the reference to minimize scope and chance of unintended issues. This
isn't the best place because it will be redefined every time through,
but this is not a heavily used code path or it would have been
found/fixed long ago.