issue2550839: Xapian, DatabaseLockError: Unable to get write lock on  db/text-index: already locked
I put in a retry loop that will attempt to get the lock. 10 tries with
a delaythatvaries from .02 seconds to .64 seconds. Total delay over 10
cycles approx. 4.5 seconds.

I can't figure out how to make pytest run two parallel processes to
test this code. I did try running:

./run_tests.py -k Xapian test/test_indexer.py &
./run_tests.py -k Xapian test/test_indexer.py

and confirmed that one of the processes seemed to hang on a test and
then threw a lock failure error. So at least the code path is being
exercised.

If anybody knows how to correctly test this I would love to do a real
test.