issue2550755: exceptions.NotFound(msg) msg is not reported to user in cgi.
When an invalid column is specified return error code 400 rather than
404. Make error code 400 also return an error message to the
user. Reported by: Bernhard Reiter.

After I implemented this it occurred to me that this may be better
implemented using a new exception BadRequest and handle that
instead. I really don't have a good feel for which is better from an
architectural view although I have a feeling the exception path would
be better.