issue2550755: exceptions.NotFound(msg) msg is not reported to user in cgi fix tests
Correct tests so they now expect NotFound exception rather than
SeriousError.

Renamed:
  test/test_cgi.py::FormTestCase::testCSVExportFailPermission
to
  test/test_cgi.py::FormTestCase::testCSVExportFailPermissionBadColumn

I think this test should return Unauthorised rather than
NotFound. Discussion is on the issue.

Also added a new test case to verify that exceptions.Unauthorised is
raised if all columns are valid and the user is not allowed to access
the class.