Issue2550934 - templating.py-indexargs_form() returns id's as space separated list not comma separated
Compare

   filter=id&id=1+2+3+4+5+6+7   (+ means space)
to
   filter=id&id=1,2,3,4,5,6,7

the latter seems to be a valid filter and the index page search will
return items with those 7 id's. The former does not return the
expected 7 items.

Check to see if the name of the field is id. If it is don't treat it
as a string which pastes all the values together with spaces.