Remove 2to3-done.txt (issue2550979).
With all the mechanical and most of the non-mechanical parts of Python
3 support done, the lists of source files in 2to3-done.txt are no
longer relevant.  The discussion of plans for Python 3 support is also
no longer relevant, with the string representations being described in
a comment in roundup/anypy/strings.py and other known outstanding
issues being described in open issues in the tracker.