Better display for Link/Multilink and content
Link/Multilink are now displayed as a dictionary by default. The format
is controlled by the setting of the @verbose option. With @verbose=0 we
get the old behavior displaying only the id. With the default @verbose=1
we get a dictionary with the id and a link inside. With @verbose=2 or
larger we get the label property in the dictionary in addition (e.g. the
name of a status or the name of a file).

The content property is also handled differently now. For @verbose < 2
we get a dictionary with a link property in it. The property points to
the standard download link for the content (or message). For
@verbose >= 2 we get the previous behavior, the content property as a
possibly very large json string.