A fix for https://issues.roundup-tracker.org/issue2551034 REST etag ordering.
Sort the items before performing the MD5 checksum. This puts the items
in the same order every time. I was able to restart the roundup-server
and get the same etag value on every restart after this patch.
Also the etag calculated using python2 and python3 are now the same.

I added a disabled test to try to validate that the etag is always the
same value. However there is no way to set the creation and activity
dates, so these fields are always different. This means the etag is
always different. If I can convince the database back end to use a
mocked date.Date.__init__ that returns the same date all the time I
can finish the test.